+++ tests/integration/test_pgvector.py
@@ -16,6 +16,18 @@ def spawn_vsb(workload, timeout=60, extra_args=None):
     return spawn_vsb_inner("pgvector", workload, timeout, extra_args, extra_env)
 
 
+# used in test_common
+def spawn_vsb_pgvector(workload, timeout=60, extra_args=None, **kwargs):
+    """Spawn an instance of pgvector vsb with the given arguments, returning the proc object,
+    its stdout and stderr.
+    """
+    extra_env = {
+        "VSB__PGVECTOR_USERNAME": "postgres",
+        "VSB__PGVECTOR_PASSWORD": "postgres",
+    }
+    return spawn_vsb_inner("pgvector", workload, timeout, extra_args, extra_env)
+
+
 class TestPgvector:
 
     def test_ivfflat(self):
